Move helpers to separate files
