+++ src/helpers/RenderButton.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const RenderButton = Component => ({
+  renderButton(editor) {
+    return <Component editor={editor} />;
+  },
+});
+
+export default RenderButton;
