+++ src/helpers/createPlugin.js
@@ -1,0 +1,3 @@
+import merge from './merge';
+
+export default merge;
