+++ src/helpers/hasBlock.js
@@ -1,0 +1,2 @@
+export default (editor, type) =>
+  editor.value.blocks.some(node => node.type === type);
