+++ src/helpers/hasWrappedBlock.js
@@ -1,0 +1,4 @@
+export default (editor, type) =>
+  editor.value.blocks.some(node =>
+    editor.value.document.getClosest(node.key, parent => parent.type === type),
+  );
