+++ src/helpers/index.js
@@ -1,21 +1,9 @@
 import React from 'react';
 
-const merge = objs => objs.reduce((r, o) => ({ ...r, ...o }), {});
-
-const createPlugin = merge;
-
-const RenderButton = Component => ({
-  renderButton(editor) {
-    return <Component editor={editor} />;
-  },
-});
-
-const hasBlock = (editor, type) =>
-  editor.value.blocks.some(node => node.type === type);
-
-const hasWrappedBlock = (editor, type) =>
-  editor.value.blocks.some(node =>
-    editor.value.document.getClosest(node.key, parent => parent.type === type),
-  );
+import merge from './merge'
+import createPlugin from './createPlugin'
+import RenderButton from './RenderButton'
+import hasBlock from './hasBlock'
+import hasWrappedBlock from './hasWrappedBlock'
 
 export { merge, createPlugin, RenderButton, hasBlock, hasWrappedBlock };
