+++ src/helpers/merge.js
@@ -1,0 +1,1 @@
+export default objs => objs.reduce((r, o) => ({ ...r, ...o }), {});
